Delete cmtrace.exe
CMTrace licensing

CMTrace is a public download from Microsoft, part of the various ConfigMgr Toolkits, but because of some utterly stupid legal terms, it’s license terms states that you actually need have a license for ConfigMgr in order to use it. To be able to read a text file, duh! But don’t worry, there is a way: According to the answer by Cathy Moya (Microsoft) in this forum thread, you can use the ConfigMgr evaluation media, on which CMTrace is available, and use it for the duration of the evaluation period :)